#92: add check whether the version is the latest
Make an HTTP call to GitHub and find out, what is the latest version of zerocracy/pages-action release.
If there is mismatch, adds a red banner on top of the generated HTML "vitals" page, saying:
```text
Since this page is rendered not by the latest version of pages-actions
plugin, some information may be rendered incorrectly.
```